Enhance purchase management: Add supplier PAN retrieval in DebitNoteController and Bill creation process. Update validation rules in BillRequest, DebitNoteRequest, and PaymentRequest to ensure dates fall within the active fiscal year. Modify Bill and DebitNote models to include supplier PAN field, improving data integrity and compliance.

+++ app/Models/BillItem.php
@@ -35,7 +35,7 @@ class BillItem extends Model
         'warehouse_id' => 'integer',
         'unit_id' => 'integer',
         'tax_id' => 'integer',
-        'quantity' => 'integer',
+        'quantity' => 'float',
         'rate' => 'float',
         'tax_amount' => 'float',
         'discount_amount' => 'float',
